database/requirements.txt
requests==2.32.3
mysql-connector-python==9.1.0
python-dotenv==1.0.1